+++ app/components/Ace/index.js
@@ -1,0 +1,1 @@
+export { default } from './Ace.svelte';
